Adding dialogParentComm to test app (#2315)
* Adding dialogParentComm to test app

* nit fix

* nit changes

* fixing e2e
