+++ apps/teams-test-app/src/components/DialogUrlAPIs.tsx
@@ -1,14 +1,12 @@
 /* eslint-disable @typescript-eslint/ban-types */
-import { dialog, DialogInfo, IAppWindow, ParentAppWindow, tasks, UrlDialogInfo } from '@microsoft/teams-js';
+import { dialog, DialogInfo, IAppWindow, tasks, UrlDialogInfo } from '@microsoft/teams-js';
 import React, { ReactElement } from 'react';
 
 import { ApiWithoutInput, ApiWithTextInput } from './utils';
 import { isTestBackCompat } from './utils/isTestBackCompat';
 import { ModuleWrapper } from './utils/ModuleWrapper';
 
-const DialogUrlAPIs = (): ReactElement => {
-  const childWindowRef = React.useRef<IAppWindow | null>(null);
-
+const DialogUrlAPIs = ({ childWindowRef }): ReactElement => {
   const openDialogHelper = (childWindow: IAppWindow, setResult: (result: string) => void): void => {
     childWindow.addEventListener('message', (message: string) => {
       // Message from parent
@@ -67,101 +65,6 @@ const DialogUrlAPIs = (): ReactElement => {
       }),
     });
 
-  const SendMessageToChild = (): ReactElement =>
-    ApiWithTextInput<string>({
-      name: 'sendMessageToChild',
-      title: 'sendMessageToChild',
-      onClick: {
-        // eslint-disable-next-line @typescript-eslint/no-empty-function
-        validateInput: () => {},
-        submit: async (message, setResult) => {
-          if (isTestBackCompat()) {
-            if (childWindowRef.current && childWindowRef.current !== null) {
-              const childWindow = childWindowRef.current;
-              const onComplete = (status: boolean, reason?: string): void => {
-                if (!status) {
-                  if (reason) {
-                    setResult(JSON.stringify(reason));
-                  } else {
-                    setResult("Status is false but there's no reason?! This shouldn't happen.");
-                  }
-                } else {
-                  setResult('Message sent to child');
-                }
-              };
-              childWindow.postMessage(message, onComplete);
-            } else {
-              setResult("childWindow doesn't exist");
-            }
-            return '';
-          } else {
-            setResult('Message sent to child');
-            dialog.url.parentCommunication.sendMessageToDialog(message);
-            return '';
-          }
-        },
-      },
-      defaultInput: '"Hello from parent"',
-    });
-
-  const SendMessageToParent = (): ReactElement =>
-    ApiWithTextInput<string>({
-      name: 'sendMessageToParent',
-      title: 'sendMessageToParent',
-      onClick: {
-        // eslint-disable-next-line @typescript-eslint/no-empty-function
-        validateInput: () => {},
-        submit: async (message, setResult) => {
-          if (isTestBackCompat()) {
-            const parentWindow = ParentAppWindow.Instance;
-            if (parentWindow) {
-              const onComplete = (status: boolean, reason?: string): void => {
-                if (!status) {
-                  if (reason) {
-                    setResult(JSON.stringify(reason));
-                  } else {
-                    setResult("Status is false but there's no reason?! This shouldn't happen.");
-                  }
-                } else {
-                  setResult('Message sent to parent');
-                }
-              };
-              parentWindow.postMessage(message, onComplete);
-            } else {
-              setResult("parentWindow doesn't exist");
-            }
-          } else {
-            setResult('Message sent to parent');
-            dialog.url.parentCommunication.sendMessageToParentFromDialog(message);
-          }
-          return '';
-        },
-      },
-      defaultInput: '"Hello from child"',
-    });
-
-  const RegisterForParentMessage = (): ReactElement =>
-    ApiWithoutInput({
-      name: 'registerForParentMessage',
-      title: 'registerForParentMessage',
-      onClick: async (setResult) => {
-        let msg = 'Completed';
-        if (isTestBackCompat()) {
-          const parentWindow = ParentAppWindow.Instance;
-          parentWindow.addEventListener('message', (message: string) => {
-            setResult(message);
-          });
-        } else {
-          const callback = (message: string): void => {
-            msg = message;
-            setResult(message);
-          };
-          dialog.url.parentCommunication.registerOnMessageFromParent(callback);
-        }
-        return msg;
-      },
-    });
-
   const SubmitDialogWithInput = (): ReactElement =>
     ApiWithTextInput<{ result?: string; appIds?: string | string[] }>({
       name: 'dialogSubmitWithInput',
@@ -193,9 +96,6 @@ const DialogUrlAPIs = (): ReactElement => {
       <CheckDialogUrlCapability />
       <OpenDialog />
       <SubmitDialogWithInput />
-      <SendMessageToChild />
-      <SendMessageToParent />
-      <RegisterForParentMessage />
     </ModuleWrapper>
   );
 };
